TT-15860, fixed upstream oauth client credentials config
